requirements.txt
Django==4.2.6
djangorestframework==3.14.0
pytils==0.4.1
django-stdimage==6.0.1
django-colorfield==0.10.1
django-imagekit==5.0.0
django-elasticsearch-dsl==8.0
django-elasticsearch-dsl-drf==0.22.5

psycopg2==2.9.9
